Migrate qos functions to ros2cli.node
