+++ ros2topic/ros2topic/verb/echo.py
@@ -27,8 +27,8 @@ from rclpy.qos_event import UnsupportedEventTypeError
 from rclpy.task import Future
 from ros2cli.node.strategy import add_arguments as add_strategy_node_arguments
 from ros2cli.node.strategy import NodeStrategy
+from ros2cli.node.qos import qos_profile_from_short_keys, add_qos_arguments
 from ros2topic.api import get_msg_class
-from ros2topic.api import qos_profile_from_short_keys, add_qos_arguments
 from ros2topic.api import TopicNameCompleter
 from ros2topic.api import unsigned_int
 from ros2topic.verb import VerbExtension
